# Conflicts:
#	src/App.css
#	src/components/Card/Card.jsx
#	src/components/HomePage.jsx
#	src/components/MainFeed.jsx
#	src/components/SideBar.jsx

+++ src/components/Todos/Todo.jsx
@@ -1,5 +1,6 @@
+import axios from "axios";
 import React, { useEffect, useState } from "react";
-import MyButton from "./MyButton"
+import ModalTodo from "./ModalTodo"
 
 
 
@@ -14,6 +15,18 @@ const Todo = props => {
 
     }, [])
 
+    const updateTodoPosted = async () => {
+        const response = await axios.put('http://localhost:8080/api/todo/id/' + props.id + '/post?isPosted=false');
+
+        console.log('isPosted false', response.data);
+
+    }
+    const deletePost = async () => {
+        const response = await axios.get('http://localhost:8080/posts/todo/' + props.id);
+        console.log(response.data);
+        await axios.delete('http://localhost:8080/posts/' + response.data.id);
+
+    }
 
 
     // Todo 업데이트
@@ -32,9 +45,20 @@ const Todo = props => {
     }
     // Todo 체크 시 삭선효과, 회색 텍스트 적용
     const checkHandler = () => {
-        setCompleted(!isCompleted);
-        props.onChecked(props.id);
+        if (isCompleted && props.posted) {
+            if (window.confirm('게시물을 삭제 하시겠습니까?')) {
+                deletePost();
+                updateTodoPosted();
+                setCompleted(!isCompleted);
+                props.onChecked(props.id);
+                setTimeout(() => { window.location.reload(); }, 1000);
+            }
 
+        } else {
+            setCompleted(!isCompleted);
+            props.onChecked(props.id);
+
+        }
 
 
     }
@@ -49,16 +73,27 @@ const Todo = props => {
         event.target.disabled = true;
     }
 
+
+
+
     // Todo 딜리트
     const deleteTodoHandler = () => {
-        props.onDelete(props.id);
+        if (props.posted) {
+            if (window.confirm('게시물을 함께 삭제 하시겠습니까?')) {
+                deletePost();
+                props.onDelete(props.id);
+                setTimeout(() => { window.location.reload(); }, 1000);
+            }
+        } else {
+            props.onDelete(props.id);
+        }
     }
 
 
     return (
         <li>
-            <div className="flex items-center w-full group">
-                <input type="checkbox" className='h-6 w-7' onChange={checkHandler} checked={isCompleted} />
+            <div className="flex items-center w-full mb-3 group">
+                <input type="checkbox" className='h-6 mr-3 w-7' onChange={checkHandler} checked={isCompleted} />
                 <div className='w-full' onDoubleClick={editableHandler}>
 
                     <input className={`p-1 h-full w-full bg-transparent ${isCompleted && 'line-through text-gray-300 ease-in duration-500'}`} // list
@@ -72,8 +107,9 @@ const Todo = props => {
                     <label htmlFor={props.id}></label>
                 </div>
 
-                <MyButton todo={title} id={props.id} todoDescription={props.todoDescription} updateSubmitHandler={updateSubmitHandler} />
-                <button className='ml-2 font-semibold cursor-default text-slate-200 hover:text-red-400' onClick={deleteTodoHandler}>X</button>
+                <ModalTodo todo={title} id={props.id} todoDescription={props.todoDescription} updateSubmitHandler={updateSubmitHandler} />
+                <img className='w-5 h-5 transition duration-300 ease-in-out delay-100 cursor-pointer hover:-translate-y-1 hover:scale-110' src="https://yestoday.s3.ap-northeast-2.amazonaws.com/trash.png" alt="" onClick={deleteTodoHandler} />
+                {/* <img className='w-5 h-5 mr-4 transition duration-300 ease-in-out delay-100 cursor-pointer hover:-translate-y-1 hover:scale-110' src="images/pencil.png" alt="" onClick={handleShow}/> */}
 
             </div>
             {/* <button onClick={()=> setModalIsOpen(true)}>Modal Open</button> */}
